+++ docker-cloud-watch.js
@@ -1,0 +1,2 @@
+require('babel-register');
+require('./lib/docker-cloud-watch');
