+++ lib/docker-cloud-watch.js
@@ -1,15 +1,15 @@
 import reloadNginx from "./reload-nginx";
-import { createStream } from "./tutum";
+import { createStream } from "./docker-cloud";
 
 //Build Initial Config
 reloadNginx();
 createStreamEvents(createStream());
 
 function createStreamEvents(stream) {
-  stream.on('open', () => console.log('Listening for Tutum Change Events'));
-  stream.on('error', (message) => console.log('Tutum Stream Error: %s', message));
+  stream.on('open', () => console.log('Listening for Docker Cloud Change Events'));
+  stream.on('error', (message) => console.log('Docker Cloud Stream Error: %s', message));
   stream.on('close', (message) => {
-    console.log('Tutum Socket connection closed unexpectedly', message);
+    console.log('Docker Cloud Socket connection closed unexpectedly', message);
     createStreamEvents(createStream());
   });
   stream.on('message', (message) => {
