+++ lib/reload-nginx.js
@@ -4,7 +4,7 @@ import { exec } from "child_process";
 import { find } from "lodash";
 
 import nginxTemplate from "./nginx-template";
-import { api as tutum } from "./tutum";
+import { api as dockerCloud } from "./docker-cloud";
 
 const configFileName = process.env.NGINX_CONFIG_FILE || "/etc/nginx/conf.d/default.conf";
 const certsPath = process.env.NGINX_CERTS || "/certs";
@@ -13,11 +13,11 @@ try {
 } catch(e) {}
 
 module.exports = function() {
-  tutum('/api/v1/container/')
+  dockerCloud('/api/app/v1/container/')
     .then((allContainers) => {
       //fetch full resource for each container
       return Promise.all(
-        allContainers.objects.map(container => tutum(container.resource_uri))
+        allContainers.objects.map(container => dockerCloud(container.resource_uri))
       );
     })
     .then(getNginxConfigs)
@@ -53,9 +53,9 @@ function getNginxConfigs(allContainers) {
   .forEach((container) => {
     const certs = find(container.container_envvars, {key: "NGINX_CERTS"});
     const allCerts = !certs ? [] : certs.value.split(",").map(val => val.split("\\n").join("\n"));
-    const ip = find(container.container_envvars, {key: "TUTUM_IP_ADDRESS"});
+    const ip = find(container.container_envvars, {key: "DOCKERCLOUD_IP_ADDRESS"});
     const port = find(container.container_envvars, {key: "NGINX_PORT"});
-    const serviceName = find(container.container_envvars, {key: "TUTUM_SERVICE_HOSTNAME"}).value;
+    const serviceName = find(container.container_envvars, {key: "DOCKERCLOUD_SERVICE_HOSTNAME"}).value;
     const virtualHosts = find(container.container_envvars, {key: "NGINX_VIRTUAL_HOST"}).value
       .split(",")
       .map((host, i) => {
