--- tutum-watch.js
@@ -1,2 +1,0 @@
-require('babel-register');
-require('./lib/tutum-watch');
